fix: create module code
