+++ ui/src/locales/zh-CN/createModule.ts
